Break SVG into multiple components
Also slight folder reshuffle
